requirements.txt
accelerate
fastapi
uvicorn
transformers
sentence-transformers
chromadb
python-dotenv
pydantic
pytest
pymorphy2
hf_xet
requests